nav controller changes

+++ app/src/main/java/za/co/woolworths/financial/services/android/geolocation/model/response/GroupSubTotal.kt
@@ -1,5 +1,5 @@
 package za.co.woolworths.financial.services.android.geolocation.model.response
 
 data class GroupSubTotal(
-    val orderTotal: Int
+    val orderTotal: Double
 )
